requirements.txt
Flask==2.0.1
pandas==1.3.2
hdbscan==0.8.27
Flask_Reuploaded==1.1.0
onnxruntime==1.8.1
Flask_Executor==0.9.4
umap_learn==0.5.1
numpy==1.20.3
flask_uploads==0.2.1
scikit_learn==0.24.2
openpyxl==3.0.7
PySide2==5.15.2
